+++ changelog.md
@@ -1,6 +1,10 @@
 # Version Log
 
 ### Version 0.3-BETA (DRAFT)
+* Added more element unlocking requirements (to be balanced)
+* Worked on Save/Load system
+  * Added Local save system (browser local storage)
+* Added Loading screen
 
 ### Version 0.2.1-BETA
 * Fixed bug: invalid crafting bond breaks production item